pyproject.toml
[project]
name = "self-pdf-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.28.0",
    "llama-index-embeddings-fastembed>=0.3.2",
    "llama-index-llms-groq>=0.3.2",
    "llama-index-llms-openai>=0.4.6",
    "llama-index-readers-file>=0.4.9",
    "llama-index-vector-stores-qdrant>=0.6.0",
]
